preliminary UDP flip mode (enable with UDP_FLIP)
 * allows "-F" with UDP "-u" in client/server mode
 * receiver will initiate "connection", i.e. send the first packets on the data channel
 * transmitter will wait for a packet and use the peer address/port
 * that way, downstream UDP throughput can be tested with clients behind NAT gateways
TODO:
currently, the receiver will do a burst of "CONN" packets to make sure the gateway
creates a connection context. There is however a chance that in congested networks,
none of them will arrive. It would be cleaner to let the transmitter give feedback
to the receiver using the reliable control channel and let the receiver do resends
in the meantime (with timeout).