check peer address after TCP accept()s and UDP connection establishments
this is to prevent man-in-the-middle attacks that could exploit
the fact that data channel connections were accepted from anywhere after
the (authenticated) control channel was established.
more specifically, a transmitter in flip/reverse mode uses the
accepted connection / determined peer address of the data channel
to send its data. an attacker (posing as the receiver) could establish
such a connection using a spoofed address and thus let an unwilling
victim receive all the traffic.